src/simulation/sumo_integration/requirements.txt
# SUMO Integration Requirements
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0

# SUMO and TraCI
traci>=1.15.0
sumo>=1.15.0

# API and networking
aiohttp>=3.8.0
requests>=2.28.0
tenacity>=8.0.0

# Data processing
plotly>=5.0.0
dash>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=0.19.0

# Logging and monitoring
loguru>=0.6.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.19.0
pytest-mock>=3.7.0

# Development
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
